Relacionamento de Aluno/Professor com Curso

+++ routes/web.php
@@ -20,19 +20,23 @@ Use App\Http\Controllers\HomeController;
 
 
 Route::get('/', [HomeController::class , 'index']);
-Route::get('/home', [HomeController::class, 'index']);
-Route::get('/user/show', [HomeController::class,'show']);
-Route::get('/user/edit/{id}', [HomeController::class,'edit']);
-Route::put('/user/update/{id}', [HomeController::class,'update']);
-Route::delete('/user/{id}', [HomeController::class,'delete']);
+Route::get('/home', [HomeController::class, 'index'])->middleware('auth');
+Route::get('/user/show', [HomeController::class,'show'])->middleware('auth');
+Route::get('/user/edit/{id}', [HomeController::class,'edit'])->middleware('auth');
+Route::put('/user/update/{id}', [HomeController::class,'update'])->middleware('auth');
+Route::delete('/user/{id}', [HomeController::class,'delete'])->middleware('auth');
 
-Route::get('/cursos', [CursoController::class , 'index']);
-Route::get('/cursos/cadastro', [CursoController::class , 'create']);
-Route::get('/cursos/{id}', [CursoController::class , 'show']);
-Route::post('/cursos', [CursoController::class, 'store']);
-Route::delete('/cursos/{id}', [CursoController::class,'delete']);
-Route::get('/cursos/edit/{id}', [CursoController::class,'edit']);
-Route::put('/cursos/update/{id}', [CursoController::class,'update']);
+Route::get('/cursos', [CursoController::class , 'index'])->middleware('auth');
+Route::get('/cursos/cadastro', [CursoController::class , 'create'])->middleware('auth');
+Route::get('/cursos/{id}', [CursoController::class , 'show'])->middleware('auth');
+Route::post('/cursos', [CursoController::class, 'store'])->middleware('auth');
+Route::delete('/cursos/{id}', [CursoController::class,'delete'])->middleware('auth');
+Route::get('/cursos/edit/{id}', [CursoController::class,'edit'])->middleware('auth');
+Route::put('/cursos/update/{id}', [CursoController::class,'update'])->middleware('auth');
+Route::put('/cursos/joinP/{id}', [CursoController::class,'InProfessor'])->middleware('auth');
+Route::post('/cursos/joinA/{id}', [CursoController::class,'InAluno'])->middleware('auth');
+Route::delete('/cursos/leaveA/{id}', [CursoController::class,'OutAluno'])->middleware('auth');
+
 
 Auth::routes();
